[KB-128] feat: CChangeRegion 컴포넌트 추가, /select-restaurant에 적용
